Refactor some FirestoreClient methods
